Push zealand version of opensds ansible script as base-code of Stor4NFV
As we discussed on last meeting, the installer script of stor4nfv
will be based on opensds ansible, so in this patch I download the
first release (zealand) of opensds code and push the ansible script
into stor4nfv repo so that we don't need to modify opensds code.
Please be free to ask if you have any question.

Change-Id: I7b50729977b195fa64e8d9a09f415d9f3329d71f
Signed-off-by: leonwang <wanghui71@huawei.com>