Add nbp-ansible script to deploy opensds northbound plugins
In this patch, I created another ansible script to deploy opensds
northbound plugins. Since we plan to integrate opensds with k8s
using flexvolume, I only add flexvolume role to work on it. But
the framework is extensible for adding more roles to integrate
with k8s in other ways.

Change-Id: Ifbd6cc38bf1789de0fcb132e6aa3cfcd9072a11b
Signed-off-by: leonwang <wanghui71@huawei.com>